HAL classes moved into different files.
